+++ src/entities/Drawable.java
@@ -1,4 +1,4 @@
-package checkingKnowledge;
+package entities;
 
 import java.awt.Graphics;
