+++ src/entities/Obstacle.java
@@ -1,4 +1,4 @@
-package checkingKnowledge;
+package entities;
 
 import java.awt.Color;
 import java.awt.Graphics;
