+++ src/entities/Road.java
@@ -1,8 +1,8 @@
-package checkingKnowledge;
+package entities;
 
-import static checkingKnowledge.Constants.GameWindow.GAME_HEIGHT;
-import static checkingKnowledge.Constants.KRoad.LEFT_LIMIT;
-import static checkingKnowledge.Constants.KRoad.RIGHT_LIMIT;
+import static utliz.Constants.GameWindow.GAME_HEIGHT;
+import static utliz.Constants.KRoad.LEFT_LIMIT;
+import static utliz.Constants.KRoad.RIGHT_LIMIT;
 
 import java.awt.Color;
 import java.awt.Graphics;
