+++ src/game/GameWindow.java
@@ -1,4 +1,4 @@
-package checkingKnowledge;
+package game;
 
 import javax.swing.JFrame;
 import javax.swing.JPanel;
