+++ src/gamestates/State.java
@@ -1,6 +1,8 @@
-package checkingKnowledge;
+package gamestates;
 
 import java.awt.Graphics;
+
+import inputs.KeyboardInterface;
 
 public interface State  extends KeyboardInterface{
