+++ src/inputs/KeyboardInterface.java
@@ -1,4 +1,4 @@
-package checkingKnowledge;
+package inputs;
 
 import java.awt.event.KeyEvent;
